fix some problem

+++ server/prisma/migrations/20260920070457_init/migration.sql
@@ -3,8 +3,8 @@ CREATE TABLE "Product" (
     "id" UUID NOT NULL,
     "DG_id" INTEGER NOT NULL,
     "name" TEXT NOT NULL,
-    "price" INTEGER NOT NULL,
-    "price_D" INTEGER NOT NULL,
+    "price" BIGINT NOT NULL,
+    "price_D" BIGINT NOT NULL,
     "image" TEXT NOT NULL,
     "inStock" BOOLEAN NOT NULL DEFAULT true,
